requirements.txt
# Python dependencies for YouTube Transcription App

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Task Queue
celery==5.3.6
redis==5.0.1

# YouTube Audio Extraction
yt-dlp==2024.1.7

# OpenAI APIs
openai==1.10.0

# Language Detection
langdetect==1.0.9

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# YAML support for config
pyyaml==6.0.1
